+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Please refer to :doc:`release notes<release_notes>`.
 - fix declaring existing model field(s) in ModelResource altering export order (#1663)
 - Updated `docker-compose` command with latest version syntax in `runtests.sh` (#1686)
 - Support export from model change form (#1687)
+- Updated Admin UI to track deleted and skipped Imports (#1691)
 
 4.0.0-beta.1 (2023-11-16)
 --------------------------