+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Please refer to :doc:`release notes<release_notes>`.
 - Updated `docker-compose` command with latest version syntax in `runtests.sh` (#1686)
 - Support export from model change form (#1687)
 - Updated Admin UI to track deleted and skipped Imports (#1691)
+- Refactored :module:`~import_export.resources` into separate modules for ``declarative`` and ``options`` (#)
 
 4.0.0-beta.1 (2023-11-16)
 --------------------------